+++ algorithm/quiescence.c
@@ -7,11 +7,11 @@
 
 #define MAX_QUIESCENCE_DEPTH 6
 
-double quiescence(BoardState *board_state,
-                  BoardStack *stack,
-                  double alpha,
-                  double beta,
-                  int depth_in_quiescence, uint64_t *nodes_searched)
+int32_t quiescence(BoardState *board_state,
+                   BoardStack *stack,
+                   int32_t alpha,
+                   int32_t beta,
+                   int depth_in_quiescence, uint64_t *nodes_searched)
 {
     (*nodes_searched)++;
     if (depth_in_quiescence >= MAX_QUIESCENCE_DEPTH)
@@ -21,8 +21,8 @@ double quiescence(BoardState *board_state,
     TT_prefetch(hash);
 
     // 1) Stand-pat
-    double stand_pat = score_board(board_state);
-    double best_score = stand_pat;
+    int32_t stand_pat = score_board(board_state);
+    int32_t best_score = stand_pat;
 
     // 2) β-cutoff on stand-pat
     if (stand_pat >= beta)
@@ -52,7 +52,7 @@ double quiescence(BoardState *board_state,
     for (uint16_t i = base; i < stack->count; i++)
     {
         BoardState *child = &stack->boards[i];
-        double score = -quiescence(child, stack, -beta, -alpha, depth_in_quiescence + 1, nodes_searched);
+        int32_t score = -quiescence(child, stack, -beta, -alpha, depth_in_quiescence + 1, nodes_searched);
 
         if (score >= beta)
         {
